+++ fast_rx_persistence/readme/usage.dart
@@ -15,13 +15,13 @@ void example() {
   // An example of persistence with value transformation
   final rx2 = 0.rx
     // The type here is [String] since that is the type stored in the store
-    ..persist<String>(
+    ..persist(
       'key',
       // The store can optionally be overridden on a per-value basis.
       // Otherwise the store passed into [FastRxPersistence.init] will be
       // used.
       store: Store(),
-      converter: InlineConverter(
+      converter: PersistenceConverter.inline(
         fromStore: int.parse,
         toStore: (value) => value.toString(),
       ),
@@ -29,6 +29,16 @@ void example() {
 
   // Saves the value to the store as a string
   rx2.value = 2;
+
+  // An example of persistence with an enum
+  final rx3 = TestEnum.one.rx
+    ..persist(
+      'key',
+      converter: EnumPersistenceConverter.string(TestEnum.values),
+    );
+
+  // Saves the value to the store as a string
+  rx3.value = TestEnum.two;
 }
 
 class Store extends FastRxPersistenceInterface {
@@ -40,3 +50,5 @@ class Store extends FastRxPersistenceInterface {
   @override
   void set(String key, Object? value) => _store[key] = value;
 }
+
+enum TestEnum { one, two }