+++ fast_rx_persistence/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.3.0
+- BREAKING: Renames `FastRxPersistenceConverter` to `PersistenceConverter`
+- BREAKING: Renames `InlineConverter` to `PersistenceConverter.inline`
+- BREAKING: Renames `EnumConverter` to `EnumPersistenceConverter`
+- BREAKING: Renames `EnumStringConverter` to `EnumPersistenceConverter.string`
+- BREAKING: Renames `EnumIntConverter` to `EnumPersistenceConverter.int`
+
+## 0.2.3
+- Lower dependency constraints on `meta` and `collection`
+
+## 0.2.2
+- Removes dependency on Flutter
+- Upgrades `fast_rx`
+
 ## 0.2.1
 - Upgrades `fast_rx`